PP-9033 Move organisation url to switch psp folder
- Reorganise `organisation URL` controllers and templates to switch-psp folder.
  Organisation URL functionality was part of KYC and Switch PSP previously and now it is only part of Switch PSP functionality
