+++ app/controllers/switch-psp/organisation-url/get.controller.test.js
@@ -27,6 +27,6 @@ describe('Org URL GET controller', () => {
   it('should render organisation URL details form', async () => {
     await getController(req, res, next)
 
-    sinon.assert.calledWith(res.render, `kyc/organisation-url`)
+    sinon.assert.calledWith(res.render, `switch-psp/organisation-url`)
   })
 })
